Funciona, pero titilan
Tengo que hacer que sea más rapido para que parezcan encendidos siempre

+++ LIBRERÍAS.X/Config_ADC.h
@@ -19,7 +19,7 @@
 // En el parámetro NUM se coloca el valor en decimal que se quiere que aparezca 
 // en el display y se carga los leds del display para que aparezca
 //******************************************************************************
-uint8_t ADC(uint8_t ADRESL_, uint8_t ADRESH_);
+int ADC(uint8_t ADRESL_, uint8_t ADRESH_);
 uint8_t SWAP_ADC(uint8_t VAL_ADC);
 uint8_t NIBBLE1_ADC(uint8_t VAL_ADC);
 uint8_t NIBBLE2_ADC(uint8_t VAL_SWAP);
